docker/bot_with_api/bot/requirements.txt
pandas==2.1.4
python-telegram-bot==20.8
requests==2.31.0